Add tomato entry point.
The tomato entry point is executed on the host,
and binds the Docker volumes with a configuration file,
environment names, or default settings.

The defaults are provided as such, on the host:

     SETTING_NAME       (    description    ) : default value

   * TOMATO_NAME        (docker's image name) : tomato
   * TOMATO_PKGDIR      (repository location) : /var/pkg/tomato
   * TOMATO_IMGDIR      (docker's image files): /usr/share/tomato
   * TOMATO_PACDIR      (pacman.d directory)  : /etc/pacman.d
   * TOMATO_MAKEPKGCONF (makepkg.conf)        : disabled by default
   * TOMATO_PKGCACHEDIR (shared pacman cache) : /var/cache/pacman/pkg

The defauls can be tuned with the --tomato-config flags,
or with an /etc/tomato.conf file or a ${HOME}/.config/tomato.conf one.

The image can be rebuilt on the fly with the --rebuild-image flag,
passed to any commands, or as is.

A shortcut is provided for the pacman users "tomato -Syu",
which will synchronize the pacman cache before updating the tomato
repository and then the host; it the same things as:

  pacman -Syuuw && tomato update [flags] && pacman -Syuu.