docker: change strategy to bind the mirrorlist.
We care about the mirrorlist only, and not all the pacman.d
directory.

If we bind a whole read-only volume /etc/pacman.d any mirrorlist
update will abort the full tomato refresh, or package installation.

pacman returns a full disk error (because it is readonly), and
abort further operations.

Let to use the same trick as for the /etc/makepkg.conf, and create
a /home/tomato/mirrorlist volume to redirect to /etc/pacman.d/mirrorlist
if the host binds /home/tomato/mirrorlist to an actual mirrorlist.

Bump version to 0.2.0 (as the volumes and logic change).