docker: fix pikaur __init__.py non user readable.
The packaging of pikaur allows only the root to read its
__init__.py file.
This prevents the tomato user, or any regular user, to call pikaur;
by design a regular user should be able invoke pikaur, the
tool itself prompt a sudo after execution.

The __ini__.py root only readable prevent that.

The pikaur root execution takes another path with a systemd dynamic user,
which we don't want within a docker image.

The tomato docker script tries to fixe the permission while building
the Docker image. Any futher pikaur update will not revert the rights
but can triger a warning telling that the permissions differs. Not an issue.