Makefile: fix common root for local and github tar
The `make tar` (or `make pkg`) didn't produce the same
tar layout that the one shipped by github (for a release).

The `make tar` didn't have a topmost "tomato-VERSION" directory,
and the PKGBUILD accepted only the `make tar` layout.

The `make tar` command and the PKGBUILD is now canonicalized
with github layout.