docker: tomato.sh filter out --noconfirm
The pacman's --confirm or pikaur's --edit flags are passed through
tomato was ignored due to the forced --noconfirm option.

Let to filter out any ${AURFLAGS} that is not compatible with
any flags passed to tomato.

For now, only the --confirm one can be used, as the --edit flag
cannot do anything as there is no $EDITOR builtin.