tomato: add a new TOMATO_EDITOR variable.
The TOMATO_EDITOR variable allows to set a custom editor within
the tomato docker image to edit PKGBUILDs.

The format is TOMATO_EDITOR=<package-name>:<command>, like:
TOMATO_EDITOR=extra/vim:/usr/bin/vim

If no TOMATO_EDITOR is set, tomato will look at the
current EDITOR environment variable, if not EDITOR is set it
will fallback to "extra/vim:/usr/bin/vim" as editor.

The package of the EDITOR is guessed from the command, as in such
way that it naivly concider the command name as the package name.
In case of trouble, explicit the editor with TOMATO_EDITOR.

The editor feature can be disabled by setting TOMATO_EDITOR=.