tomato: try to be less anoying about the EDITOR=
Before to emit a warning about the EDITOR= not be the same as the one
configured for tomato, apply a which on non relative set.

I we had EDITOR=vim and TOMATO_EDITOR=/usr/bin/vim, tomato
complained that the EDITOR= does not yet reflect the value within the
tomato image, even if it is actually the same.