docker: fix unseen issue with tomato list all
The check test to avoid `repose` error about non existing tomato
database was wrong if the current dir was not properly set.
This worked but no more. Let to use a full path.

The method to check the avaiable packages was used by others, hence some
methods failed to report correctly the informations, such as
`tomato list status`.