docker: use the archlinux:base-devel image.
The archlinux/base is no more available (#2) and was moved to
an offical Docker image as `archlinux`. Two flavors are provided:

  - latest, base
  - base-devel

Plus some pinned versions.

As tomato is intented to be used to build the AUR's packages,
the image uses the base-devel flavor.

Note: No more cleanup hook script is provided (#3).