docker: fix sudo requires a password for pikaur.
Pikaur changed is `sudo` handling, and tries to jump to
sudo itself if not root; it seems.

Let to add another suders files to allow any users to
call `/usr/bin/pikaur` as any other user.

In our side we are making sure to not call pikaur as
root, and pikaur make sure be called itself somehow
with `sudo`.

Anyway, we can't call pkiaur as root even if we wanted to,
as the dynamic user feature in such case requires a
systemd tooling not available in the docker image (as systemd
as no power here).