tomato: add a TOMATO_ULIMIT conf option.
`fakeroot` may hang inside docker, to prevent that an `--ulimit nofile=`
should be set.

The default value of `DOCKER_ULIMIT` is `nofile=1024:524288`; value
found elsewhere as resolution.

To override this value, the `DOCKER_ULIMIT` can be set, or blanked,
from the tomato.conf file.

If blank, no ulimit option is passed to docker during the build or
during the image run.